feat: add reusable scrollable list
